Moves album list fetching for a user to profiles app
