+++ moments/templates/albums/album_detail.html
@@ -1,15 +1,15 @@
 {% extends "base.html" %}
 
 {% block content %}
-<h1 class="display-4">{{ album.name }}</h1>
- <div class="card-columns">
-    {% for photo in album.photo_set.all %}
-        <div class="card">
-            <div class="card-body">
+    <h1 class="display-4">{{ album.name }}</h1>
+    <div class="card-columns">
+        {% for photo in album.photo_set.all %}
+            <div class="card">
                 <img class="card-img-top" src="{{ photo.image.url }}" alt="{{ photo.title }}">
-                <h5 class="card-title">{{ photo.title }}</h5>
+                <div class="card-body">
+                    <h5 class="card-title">{{ photo.title }}</h5>
+                </div>
             </div>
-        </div>
-    {% endfor %}
-</div>
+        {% endfor %}
+    </div>
 {% endblock %}
